Remove ElementTrait and add SearchTrait [skip ci]
