+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphAccordion.php
@@ -6,9 +6,8 @@ namespace Drupal\server_general\Plugin\EntityViewBuilder;
 
 use Drupal\paragraphs\ParagraphInterface;
 use Drupal\pluggable_entity_view_builder\EntityViewBuilderPluginAbstract;
-use Drupal\server_general\AccordionTrait;
 use Drupal\server_general\ElementLayoutTrait;
-use Drupal\server_general\ElementTrait;
+use Drupal\server_general\ElementTrait\AccordionTrait;
 use Drupal\server_general\ElementWrapTrait;
 use Drupal\server_general\ProcessedTextBuilderTrait;
 
@@ -26,7 +25,6 @@ use Drupal\server_general\ProcessedTextBuilderTrait;
 class ParagraphAccordion extends EntityViewBuilderPluginAbstract {
 
   use AccordionTrait;
-  use ElementTrait;
   use ElementLayoutTrait;
   use ElementWrapTrait;
   use ProcessedTextBuilderTrait;
