+++ web/modules/custom/server_general/src/Plugin/EntityViewBuilder/ParagraphRelatedContent.php
@@ -21,7 +21,7 @@ class ParagraphRelatedContent extends EntityViewBuilderPluginAbstract {
 
   use ButtonTrait;
   use ElementTrait\CarouselTrait;
-  use ElementTrait;
+
   use ProcessedTextBuilderTrait;
 
   /**
