+++ web/modules/custom/server_style_guide/src/Controller/StyleGuideController.php
@@ -11,7 +11,17 @@ use Drupal\server_general\ButtonTrait;
 use Drupal\server_general\ElementLayoutTrait;
 use Drupal\server_general\ElementMediaTrait;
 use Drupal\server_general\ElementNodeNewsTrait;
-use Drupal\server_general\ElementTrait;
+use Drupal\server_general\ElementTrait\CardTrait;
+use Drupal\server_general\ElementTrait\CarouselTrait;
+use Drupal\server_general\ElementTrait\CtaTrait;
+use Drupal\server_general\ElementTrait\DocumentsTrait;
+use Drupal\server_general\ElementTrait\HeroTrait;
+use Drupal\server_general\ElementTrait\InfoCardTrait;
+use Drupal\server_general\ElementTrait\NewsTeasersTrait;
+use Drupal\server_general\ElementTrait\PeopleTeasersTrait;
+use Drupal\server_general\ElementTrait\QuickLinksTrait;
+use Drupal\server_general\ElementTrait\QuoteTrait;
+use Drupal\server_general\ElementTrait\SearchTrait;
 use Drupal\server_general\ElementWrapTrait;
 use Drupal\server_general\InnerElementTrait;
 use Drupal\server_general\LinkTrait;
@@ -29,23 +39,23 @@ class StyleGuideController extends ControllerBase {
   use AccordionTrait;
   use BuildFieldTrait;
   use ButtonTrait;
-  use ElementTrait\CarouselTrait;
-  use ElementTrait\CtaTrait;
-  use ElementTrait\HeroTrait;
-  use ElementTrait\CardTrait;
-  use ElementTrait\DocumentsTrait;
-  use ElementTrait\InfoCardTrait;
-  use ElementTrait\PeopleTeasersTrait;
-  use ElementTrait\QuickLinksTrait;
-  use ElementTrait\QuoteTrait;
+  use CardTrait;
+  use CarouselTrait;
+  use CtaTrait;
+  use DocumentsTrait;
   use ElementMediaTrait;
   use ElementNodeNewsTrait;
-  use ElementTrait;
   use ElementLayoutTrait;
   use ElementWrapTrait;
+  use HeroTrait;
+  use InfoCardTrait;
   use InnerElementTrait;
   use LinkTrait;
-  use ElementTrait\NewsTeasersTrait;
+  use NewsTeasersTrait;
+  use PeopleTeasersTrait;
+  use QuickLinksTrait;
+  use QuoteTrait;
+  use SearchTrait;
   use SocialShareTrait;
   use StyleGuideElementWrapTrait;
   use TagTrait;
